Rename: 변경된 폴더명으로 변경
